Replace my custom vanilla emacs by doom emacs
Instead of customizing emacs from scratch, we now
rely on doom emacs for most of the stuff.

Generally, almost every piece of functionality should
still be available, except that some keybindings have
changed (mostly to doom emacs's default bindings which
I found quite ok), completion from ivy to vertico, and
a few more changes which should be functional-wise iso
to what we had previously.

Moreover, we went back to 27.x (without native-comp)
for now.

Lastly, my old vanilla config is kept, but shall
be removed in a coming patch as soon as we stabilize
doom.